+++ tests/integration/test_end_to_end.py
@@ -90,4 +90,4 @@ def test_enable_compression():
 
 
 def test_version():
-    assert tp.__version__ == "0.6.0"
+    assert tp.__version__ == "0.7.0"